Remove slime overlay
This is horrendously broken, as I've just discovered to my detriment after mining out a double "slime chunk" only for it to spawn zero slimes... because neither is actually a slime chunk.

Tell-tale signs of it being completely and utterly broken include it marking different chunks as slime chunks on different map orientations.